backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
gunicorn>=21.0,<22.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
django-cors-headers>=4.3,<5.0
django-allauth>=0.61,<1.0
social-auth-app-django>=5.4,<6.0
requests>=2.31,<3.0
